+++ src/test/resources/com/pestphp/pest/inspections/ExpectCallsWithOtherStatementsBetween.php
@@ -1,0 +1,5 @@
+<?php
+
+expect($answer)->toBe(42);
+Earth::tryCalculateQuestion();
+expect($question)->toBeNull();