+++ src/test/resources/com/pestphp/pest/inspections/ManyExpectCall.after.php
@@ -1,0 +1,9 @@
+<?php
+
+expect($answer)->toBe(42)
+    ->and($question)->toBeNull()
+    ->and($question)->toBeNull()
+    ->and($question)->toBeNull()
+    ->and($question)->toBeNull()
+    ->and($question)->toBeNull()
+    ->and($question)->toBeNull();