+++ src/test/resources/com/pestphp/pest/inspections/SingleExpectCall.php
@@ -1,0 +1,3 @@
+<?php
+
+expect($answer)->toBe(42);